+++ packages/node/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fallback to singular provider if batch provider is not supported (#144)
 
 ### Fixed
 - Fix missing ds option for event in dictionary query entries (#145)