+++ packages/node/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Moved `CosmosChainType` and `CosmosProjectNetConfig` to `@subql/common-cosmos` (#160)
 
 ## [2.10.1] - 2023-08-03
 ### Fixed